Progress removing page and line numbers
This is how far I've gotten. Basically, everything between [ and ] (and the brackets themselves) needs to be removed in the txt documents, as they are only used for indicating page and line numbers. This file is a test on a string, I have not managed to do it on a file yet.