Add files via upload
The 'OldFormat_Precleaning_Arma3.csv' file has been uploaded to serve as an example what the CSV files looked like before I cleaned them up for use. After splitting up the files according to the user rating, it became clear that in order for the CSV file to be parsed, I had to remove all other columns, as the variation in punctuation used by Steam users made it difficult to work with a delimiter. I tried using different encoding types, using delimiters such as the pipe (  |  ), but nothing really appeared to work except for stripping out all the irrelevant columns.